Added Server Test unit
