+++ package.json
@@ -32,6 +32,7 @@
     "@vitest/ui": "^1.6.0",
     "jsdom": "^24.0.0",
     "nodemon": "^3.1.2",
+    "supertest": "^7.0.0",
     "typescript": "^5.0.4",
     "vite": "^4.2.1",
     "vitest": "^1.6.0"
