+++ src/test/main.test.jsx
@@ -5,6 +5,5 @@ import React from "react";
 describe('main file', () => {
     it('should render React', () => {
         render(<React.StrictMode/>);
-
     })
 })
